Don't crash in progressbar when tty width is zero
Adam Miskiewicz reported a RangeError raised when running on
macOS 10.12, with Node 8.6.  This would would have been caused by
process.stdout.columns being zero.

In cases where the width is unknown, skip clearing the previous by
overwriting with spaces, and leave for the next output to overwrite.

This should be sufficient, as the progress bar collapses to a short
output such as "[] 836/836".