ci(circle): Fix deploy job so it only runs on version tags
**Summary**

This PR fixes the deploy job to only run for version tags. It also renames it to "Publish" to convey
the actual task being performed since this job only publishes to NPM. Everything else is handled by
our webhooks.

Source: https://circleci.com/docs/2.0/workflows/#git-tag-job-execution

**Test plan**

Builds on master should be fixed and when something is tagged, it should be deployed to NPM.