ci(lint): Fix the lint command so it doesn't run dupe-check yet (#4949)
**Summary**

This is a follow up to #3929. The way we run `dupe-check` was not compatible with Windows builds and it also didn't ignore any duplications in lint check. This PR removes it from the lint stage until we don't have any duplication warnings. It also fixes the way js-inspect is called so it is Windows-compatible.

**Test plan**

Builds should pass both on AppVeyor and CircleCI. `yarn lint` should also finish without errors and `yarn dupe-check` should exit with an error and a list of code duplications in the repo.