fix(request): Fix https_proxy env var (#4903)
**Summary**

Fixes #4885. In #4761 I started always passing the proxy to the `request` library to prevent it from falling back to env vars HTTPS_PROXY and HTTP_PROXY. At the time I thought that Yarn would pick up these env vars through its config system and pass them down itself.

It turns out this was not true and caused #4885. Now handling an explicit `false` in the request-manager to allow it to override the `https_proxy` if set. Otherwise, resume allowing request library to pick up the env vars.

**Test plan**

Manually tested. Since the actual requests are proxies under unit test, this is difficult to unit test.