fix(install): use node-gyp from homebrew npm
**Summary**

With this yarn will be able to discover and use the node-gyp from the
homebrew installed npm on macOS instead of falling back to globally
installing node-gyp every time a native addon needs to be compiled from
source.

Homebrew installs a clean copy of npm inside a libexec folder together
with node.
Previously yarn didn't look there when trying to locate node-gyp and the
globally install node-gyp fallback would be used every time when
building native addons with a yarn version from homebrew.
This PR adds the libexec path of node-gyp from homebrew to the node-gyp
search paths of yarn, making it possible to compile native addons using
the homebrew npm provided node-gyp with yarn without relying on this
fallback.

**Test plan**

This can't be tested outside a homebrew environment.

A way to manually test this on macOS, by installing a test build of this
PR with homebrew and trying to install a native addon with it, is
provided in the PR description.