ci(circle): Fix cache key setup for proper node_modules sharing
**Summary**

We had a tiered cache key setup for some reason (probably remnant of the pre-macOS builds config) which was breaking macOS builds when a new dpenednecy was introduced due to common install was done on a Docker machine and acached with a key including the architecture. This patch changes that and ties everything to a single cache key.

**Test plan**

CircleCI builds should pass without issues.