test(resolution): Added a (failing) unit test to reproduce issue #4743
Added a failing unit test to reproduce issue #4743. It seems that if a peerDep exists deeper in the
dep tree than where it is included, yarn will output a earning, even if that peerDep is satisfied by
the same library included shallower in the tree, or at the root level.