test(integration): Fix failing react-scripts test due to unexpected warning
**Summary**

We don't expect any warnings when installing high-profile and well-tested packages like
react-scripts so we have an integration test that enforces 0 warnings. Unfortunately a package got renamed and started producing a warning about this so the tests started failing. This PR relaxes the test to only fail on peer dependency warnings.

**Test plan**

All builds and tests should pass.