fix(resolution): No longer warn for mising peerDep when it exists at multiple levels.
A missing peerDep warning was being issued if the exact same pattern was a deep transitive dep and a
direct dep. This was due to only the first request for a pattern being added to the list of requests
that peerDep was checking. Now all references are tracked. Also fixed a logic error in Warn where a
dep would be reported multiple times.

#4743