fix(fetching): Moved "link:" specific logic to new LinkFetcher. Fix bug with bin links for "link:" t
Previously there was logic specific to the "link:" type in PackageFetcher. This violates
single-responsibity and open-closed principles. This was moved to a new LinkFetcher to align with
the other *Fetcher classes that house logic specific to dep types. There was also a bug where no
Manifest (package.json) was loaded for "link:" deps, which led to bin links not being created,
engines ignored, etc. The Manifest is now loaded from the linked dependency if it exists, so these
bugs no longer happen.