fix(cli): check if workspaces setting in package.json is an array (#5089) (#5114)
Summary

Display a helpful error message if workspaces in package.json is not an array. Currently just displays: error An unexpected error occurred: "patterns.map is not a function".

Fixes #5089

Test plan

Create a package.json file with workspaces: "packages/*"
Run yarn and observe that this error message is shown: The workspaces setting in package.json must be an array.