fix(resolution) Eliminate "missing peerDep" warning when dep exists at root level. (#5088)
* test(resolution): Added a (failing) unit test to reproduce issue #4743

Added a failing unit test to reproduce issue #4743. It seems that if a peerDep exists deeper in the
dep tree than where it is included, yarn will output a earning, even if that peerDep is satisfied by
the same library included shallower in the tree, or at the root level.

* fix(resolution): No longer warn for mising peerDep when it exists at multiple levels.

A missing peerDep warning was being issued if the exact same pattern was a deep transitive dep and a
direct dep. This was due to only the first request for a pattern being added to the list of requests
that peerDep was checking. Now all references are tracked. Also fixed a logic error in Warn where a
dep would be reported multiple times.

#4743

* fix(tests): Remove snapshot test of "yarn why" due to different output local vs ci. Now assert with

Remove snapshot test of "yarn why" due to different output local vs ci. Now assert with actual
object value comparison.