fix(cli): adjust colors for unstable version updates
This change special-cases the color indicator for updates to
*unstable* package versions, to more accurately indicate
backwards compatibility.

For 0.x.y versions, increments to the *minor* version number will
be considered a *major* version change, and increments to the
*patch* number, will be considered a *minor* version change.

For 0.0.x versions, increments to *any* version number will be
considered a *major* version change.

This matches the behavior of how yarn & npm treats updates to
packages defined with a caret-range.