fix(upgrade): No longer warn when upgrading a dev or optional dep
Previously the upgrade command would call add with the list of packages to upgrade to. No "--dev"
flag would be carried over and this list would contain all dependencies. As a result, Add would
report a warning `{package} is already in "devDependencies". Please remove existing entry first
before adding it to "dependencies".`