fix(upgrade): Fix for pattern not working with upgrade (#5662)
fixes #5660

**Summary**

Before it needed to have package argument to enter the block to filter the deps. Now it will do it if either a package is given or a --pattern is given.

**Test plan**

Corrected my incorrectly created tests.
Also `is-negative-zero` would not update as there is no new version which would be found in a caret. So changed that to `array-union` which has a version `1.0.2`