fix(run): Remove notice and licenseText fields from npm_package_* (#5666)
**Summary**

Fixes #5420. NOTICE and licenseText for packages can be quite long which
can cause E2BIG errors when trying to put it into environment variables
when running scripts. Since these fields are not useful for scripts
anyway, add them to the current blacklist which already has a similar
field in it, `readme`.

**Test plan**

Added an automated test to ensure the env doesn't have these fields but
keeps any other keys.