fix(publish): Fixes the path parameter when specifying a dir to publish.
When specifying a directory as the first parameter to `publish` yarn was ignoring it and just using
`config.cwd`. This fix checks to see if the passed parameter is to a directory and if it is it sets
`config.cwd` to that directory, so that manifest loading and lifecycle script execution occur
correctly in that specified directory.

fixes #5678