fix(publish): Run pre and postpack lifecycle scripts on publish (#5712)
**Summary**

npm runs the pre and post pack scripts on publish, but yarn was not. Added these scripts.

fixes #5707

**Test plan**

Added a test that regex's the output to look for scripts being run.
If you look at the test, the regex is kinda weird, but I did it that way to make sure the log statements were in the correct order in the output.