chore(deps): update deprecated deps for node 10 (#5763)
**Summary**
These are a few upgrades of dependencies in order to address issues on node 10 (in relation to: https://github.com/yarnpkg/yarn/issues/5477)

*duplexify* *3.5.0* => *3.5.4* (https://github.com/mafintosh/duplexify/commit/00d08fa3f223e1c7d1c05e0350c236d1c4e32d09)
*v8-compile-cache* *^1.1.0* => *^2.0.0* (https://github.com/zertosh/v8-compile-cache/pull/7)
*natives* *1.1.0* => *1.1.3* (https://github.com/gulpjs/gulp/issues/2162)

The last did not have to do with `new Buffer` but was something I personally encountered when using `gulp` for building in `node 10`. I don't know how common this is, but the fix is easy and within scope, so decided to include it.


**Test plan**

CI should be green.