fix(run): Fix custom shell execution
A previous change removed the -c option that was passed to the shell, causing script like "echo foo"
to fail if script-shell was set to bash. This change cleans up the way custom shell is specified,
passing the shell to node's child process spawner instead of trying to make it part of our command
string. Doing it this way seems to account for the previously required -c flag.

fixes #5699