Support MSYS{2} bash in bin/yarn
The path translation happens differently, or perhaps conditionally dependdding on whether msys detects
the target binary is a native Win32 application or not. While Git Bash seems to treat its `winpty` as
a Win32 binary, the version of `winpty` which is compatible with MSYS2 does not trigger the path conversion.
So, add a case for it to first run `cygpath` on the base dir.

To test, `yarn` executed from within bash within MSYS2 correct executes, instead of failing with a path
related error. `yarn` continues to execute correctly from within Git Bash.