fix(resolution): Handle git urls that have a hash in the branch name.
Previously yarn was using a string split('#') to separate the URL from the branch name. If a branch
also had a hash in it, then the rest of the split chunks were ignored, meaning only the chunk before
the first '#' was used as the branch name. This change handles those URLs correctly, preserving the
'#'s in URL hashes.

#5954