fix(resolution): Improve dependency resolution performance
Instead of continuing to re-resolve a package in the dependency tree that has already been resolved,
add a request to the existing references requests array. This speeds up very deep/complex dep trees
as seen in workspaces where the workspace modules all depend on each other. Previously the building
of the tree was leading to out of memory errors or just never finishing.

fixes #5726 #5628