Ensure the useful output is displayed in silent mode
For `list` and `licenses list` commands. They make no sense as commands
without the tree output, but it does make sense to use --silent to
suppress informational output about yarn version and how long the
commands took if you are redirecting to a file, for example to compare
with other projects or other versions of the same project.