Improve performance of selecting packages in step 5 of install
See https://github.com/yarnpkg/yarn/issues/6105 for motivation.

This approach avoids repeated CPU-intensive operations on packages
that are blocked in the queue by 1) noting which are blocked on
packages currently being installed and 2) integrating our checks
for cases where we need to break cycles with our search for a
package that is ready to install.