test(warning): Jest: Ignore build directory modules to avoid module name collision (#6252)
The test suite displays the following warning message when run:
```
jest-haste-map: @providesModule naming collision:
  Duplicate module name: yarn
  Paths: <repo directory>\dist\package.json collides with
    <repo directory>\package.json
```

This is because the Jest module system finds two copies of Yarn; the
primary package and the build output. This causes ambiguity when loading
`yarn` from a test, or a module that depends upon yarn.

Adding the build directory to `modulePathIgnorePatterns` prevents this
message. This means that if `yarn` is imported from within a test (which
it is not currently), the version found will be the primary package, not
the built version.

This seems acceptable. Currently, components of `yarn` are referenced by
filename rather than by package name, and they are all from the `src`
directory. The many cases where the built `yarn` is run in tests are all
done through a child process, where this change would not be applicable
anyway.