Fix finding root workspace when the pattern ends with a slash
If a workspace pattern finishes with a traling slash, like in

    workspaces: [
      "package-b/"
    ]

then yarn will fail to detect the root workspace directory, thinking
that the package is not part of a workspace.

However, yarn will still list the package with `yarn workspaces info`
for example.

To solve this inconsistency, the relative path will be match against
the patterns with and without trailing slash.