Add the string '@generated' to the lockfile
Various tools such as linters look for the string '@generated' in a file
and use this to skip processing of the file. This change will help prevent
linters from targeting the yarn.lock lockfile, as there is no benefit to
performing things such as line length checks on this file.