fix(upgrade): Fix yarn upgrade when using --scope with exotic dependencies.
Exotic dependencies (github urls for example) were not being unlocked from the lockfile when running
`yarn upgrade --scope @whatever/` which was resulting in dependencies not being upgraded. This was
due to trying to find the dep in the lockfile just by it's url instead of in `name@url` format.

fixes #7016