fix(upgrade) Fix basicSemverOperatorRegex (#7080)
**Summary**

Fix bug #7079 of `basicSemverOperatorRegex`. It doesn't match the range operator correctly if the dependency range is specified with `>=`. For example, when running `yarn upgrade --latest`, if a dependency is specified with `>=`, the matched result will be `>`, which is not correct. 

This is because the regular expression engine looks for alternations one-by-one. We can fix it by putting `>=` before `>`.

fixes #7079