Resolve .bin path based on --modules-folder
The first entry in registryFolders array corresponds to the location
provided with `--modules-folder` to specify where to store node_modules.
This currently does a `path.join` on the current location and the
specified directory which prevents the modules folder from existing
outside of the current tree.

For example, if we run:

```
yarn --modules-folder install
```

we will successfully get the modules installed into /tmp/node_modules,
but when we later run

```
yarn --modules-folder /tmp/node_modules
```

we will get `/root/workspace/tmp/node_modules` instead of
`/tmp/node_modules` as we expect.

This also happens when injecting the .bin directory into the PATH
variable when executing lifecycle scripts.