Include status message (reason phrase) in error message
When an HTTP request fails, the status code and description is logged/output. That description currently is derived by mapping the status code to a predefined mapping of messages. This change uses the HTTP response status line reason phrase if it is available, falling back to the predefined mapping to a message if the server didn't provide one.

The reason phrase that the server responds with may have more information that will be helpful for the user. For example, Sonatype Nexus Firewall includes a message in the reason phrase indicating the requested package is blocked. Other software similarly returns useful information in the reason phrase.

See:
* https://www.w3.org/Protocols/rfc2616/rfc2616-sec6.html
* https://nodejs.org/api/http.html#messagestatusmessage